refactor: 셀프 클로징 적용 및 폴더 구조 변경
